+++ EppoEngine/Source/Platform/Vulkan/VulkanContext.h
@@ -1,7 +1,12 @@
 #pragma once
 
+#include "Platform/Vulkan/LogicalDevice.h"
+#include "Platform/Vulkan/PhysicalDevice.h"
+#include "Platform/Vulkan/Swapchain.h"
 #include "Platform/Vulkan/Vulkan.h"
 #include "Renderer/RendererContext.h"
+
+#include <deque>
 
 struct GLFWwindow;
 
@@ -16,12 +21,24 @@ namespace Eppo
 		virtual void Init() override;
 		virtual void Shutdown() override;
 
+		void SubmitResourceFree(std::function<void()> fn);
+
+	private:
+		bool HasValidationSupport() const;
+		std::vector<const char*> GetRequiredExtensions() const;
+
 	private:
 		GLFWwindow* m_WindowHandle;
 
 		VkInstance m_VulkanInstance;
 		VkDebugUtilsMessengerEXT m_DebugMessenger;
 
+		Ref<PhysicalDevice> m_PhysicalDevice;
+		Ref<LogicalDevice> m_LogicalDevice;
+		Ref<Swapchain> m_Swapchain;
 
+		// Vulkan resource management
+		std::deque<std::function<void()>> m_ResourceFreeCommands;
+		uint32_t m_ResourceFreeCommandCount = 0;
 	};
 }
