+++ EppoEngine/Source/Renderer/Framebuffer.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "Renderer/Vulkan.h"
 #include "Renderer/Image.h"
 
 namespace Eppo
@@ -27,40 +26,27 @@ namespace Eppo
 	class Framebuffer
 	{
 	public:
-		Framebuffer(const FramebufferSpecification& specification);
-		~Framebuffer();
+		virtual ~Framebuffer() {};
 
-		void Create();
-		void Cleanup();
 		virtual void Resize(uint32_t width, uint32_t height) = 0;
 
 		const FramebufferSpecification& GetSpecification() const { return m_Specification; }
 
 		Ref<Image> GetFinalImage() { return m_ImageAttachments[0]; };
 
-		VkFramebuffer GetFramebuffer() const { return m_Framebuffer; }
-		VkRenderPass GetRenderPass() const { return m_RenderPass; }
-
 		uint32_t GetWidth() const { return m_Specification.Width; }
 		uint32_t GetHeight() const { return m_Specification.Height; }
-		VkExtent2D GetExtent() const { return { GetWidth(), GetHeight() }; }
-
-		const std::vector<VkClearValue>& GetClearValues() const { return m_ClearValues; }
-
+		
 		bool HasDepthAttachment() const { return m_DepthTesting; }
 		Ref<Image> GetDepthImage() { return m_DepthImage; }
 
 		static Ref<Framebuffer> Create(const FramebufferSpecification& specification);
 
-	private:
+	protected:
 		FramebufferSpecification m_Specification;
 
-		VkFramebuffer m_Framebuffer;
-		VkRenderPass m_RenderPass;
 		std::vector<Ref<Image>> m_ImageAttachments;
 		Ref<Image> m_DepthImage;
-
-		std::vector<VkClearValue> m_ClearValues;
 
 		bool m_DepthTesting = false;
 	};
