+++ EppoEngine/Source/Renderer/RendererContext.cpp
@@ -27,158 +27,32 @@ namespace Eppo
 		}
 	}
 
-	void RendererContext::Init()
-	{
-		EPPO_PROFILE_FUNCTION("RendererContext::Init");
+	//void RendererContext::Init()
+	//{
+	//	// Profiler
+	//	m_TracyContexts.resize(VulkanConfig::MaxFramesInFlight);
+	//	for (uint32_t i = 0; i < VulkanConfig::MaxFramesInFlight; i++)
+	//		m_TracyContexts[i] = TracyVkContext(m_PhysicalDevice->GetNativeDevice(), m_LogicalDevice->GetNativeDevice(), m_LogicalDevice->GetGraphicsQueue(), m_Swapchain->m_CommandBuffers[i]);
 
-		// Vulkan instance
-		if (VulkanConfig::EnableValidation)
-			EPPO_ASSERT(HasValidationSupport());
+	//	SubmitResourceFree([=]()
+	//	{
+	//		for (uint32_t i = 0; i < VulkanConfig::MaxFramesInFlight; i++)
+	//			TracyVkDestroy(m_TracyContexts[i]);
+	//	});
+	//}
 
-		VkApplicationInfo appInfo{};
-		appInfo.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO;
-		appInfo.pApplicationName = "EppoEngine";
-		appInfo.applicationVersion = VK_MAKE_VERSION(1, 0, 0);
-		appInfo.pEngineName = "EppoEngine";
-		appInfo.engineVersion = VK_MAKE_VERSION(1, 0, 0);
-		appInfo.apiVersion = VK_API_VERSION_1_3;
+	//void RendererContext::WaitIdle()
+	//{
+	//	EPPO_PROFILE_FUNCTION("RendererContext::WaitIdle");
 
-		auto extensions = GetRequiredExtensions();
-
-		VkInstanceCreateInfo createInfo{};
-		createInfo.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO;
-		createInfo.pApplicationInfo = &appInfo;
-		createInfo.enabledExtensionCount = (uint32_t)extensions.size();
-		createInfo.ppEnabledExtensionNames = extensions.data();
-
-		VkDebugUtilsMessengerCreateInfoEXT debugCreateInfo{};
-		if (VulkanConfig::EnableValidation)
-		{
-			debugCreateInfo.sType = VK_STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CREATE_INFO_EXT;
-			debugCreateInfo.messageSeverity = VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT;
-			debugCreateInfo.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT;
-			debugCreateInfo.pfnUserCallback = DebugCallback;
-			debugCreateInfo.pUserData = nullptr;
-
-			createInfo.enabledLayerCount = (uint32_t)VulkanConfig::ValidationLayers.size();
-			createInfo.ppEnabledLayerNames = VulkanConfig::ValidationLayers.data();
-			createInfo.pNext = (VkDebugUtilsMessengerCreateInfoEXT*)&debugCreateInfo;
-		}
-		else
-		{
-			createInfo.enabledLayerCount = 0;
-			createInfo.pNext = nullptr;
-		}
-
-		VK_CHECK(vkCreateInstance(&createInfo, nullptr, &m_VulkanInstance), "Failed to create instance!");
-
-		// Debug messenger
-		if (VulkanConfig::EnableValidation)
-			VK_CHECK(CreateDebugUtilsMessengerEXT(m_VulkanInstance, &debugCreateInfo, nullptr, &m_DebugMessenger), "Failed to create debug messenger!");
-
-		// Devices
-		m_PhysicalDevice = CreateRef<PhysicalDevice>();
-		m_LogicalDevice = CreateRef<LogicalDevice>(m_PhysicalDevice);
-
-		// Allocator
-		Allocator::Init();
-
-		// Swapchain
-		m_Swapchain = CreateRef<Swapchain>();
-		
-		// Profiler
-		m_TracyContexts.resize(VulkanConfig::MaxFramesInFlight);
-		for (uint32_t i = 0; i < VulkanConfig::MaxFramesInFlight; i++)
-			m_TracyContexts[i] = TracyVkContext(m_PhysicalDevice->GetNativeDevice(), m_LogicalDevice->GetNativeDevice(), m_LogicalDevice->GetGraphicsQueue(), m_Swapchain->m_CommandBuffers[i]);
-
-		SubmitResourceFree([=]()
-		{
-			for (uint32_t i = 0; i < VulkanConfig::MaxFramesInFlight; i++)
-				TracyVkDestroy(m_TracyContexts[i]);
-		});
-	}
-
-	void RendererContext::Shutdown()
-	{
-		EPPO_PROFILE_FUNCTION("RendererContext::Shutdown");
-
-		for (uint32_t i = 0; i < m_ResourceFreeCommandCount; i++)
-		{
-			m_ResourceFreeCommands.back()();
-			m_ResourceFreeCommands.pop_back();
-		}
-
-		if (VulkanConfig::EnableValidation)
-			DestroyDebugUtilsMessengerEXT(m_VulkanInstance, m_DebugMessenger, nullptr);
-	
-		vkDestroyInstance(m_VulkanInstance, nullptr);
-	}
-
-	void RendererContext::WaitIdle()
-	{
-		EPPO_PROFILE_FUNCTION("RendererContext::WaitIdle");
-
-		VkDevice device = m_LogicalDevice->GetNativeDevice();
-		vkDeviceWaitIdle(device);
-	}
-
-	void RendererContext::SubmitResourceFree(std::function<void()> fn)
-	{
-		EPPO_PROFILE_FUNCTION("RendererContext::SubmitResourceFree");
-
-		m_ResourceFreeCommands.push_back(fn);
-		m_ResourceFreeCommandCount++;
-	}
+	//	VkDevice device = m_LogicalDevice->GetNativeDevice();
+	//	vkDeviceWaitIdle(device);
+	//}
 
 	Ref<RendererContext> RendererContext::Get()
 	{
 		EPPO_PROFILE_FUNCTION("RendererContext::Get");
 
 		return Application::Get().GetWindow().GetRendererContext();
-	}
-
-	bool RendererContext::HasValidationSupport()
-	{
-		EPPO_PROFILE_FUNCTION("RendererContext::HasValidationSupport");
-
-		uint32_t layerCount = 0;
-		vkEnumerateInstanceLayerProperties(&layerCount, nullptr);
-
-		std::vector<VkLayerProperties> availableLayers(layerCount);
-		vkEnumerateInstanceLayerProperties(&layerCount, availableLayers.data());
-
-		for (const char* layerName : VulkanConfig::ValidationLayers)
-		{
-			bool layerFound{ false };
-
-			for (const auto& layerProperties : availableLayers)
-			{
-				if (strcmp(layerName, layerProperties.layerName) == 0)
-				{
-					layerFound = true;
-					break;
-				}
-			}
-
-			if (!layerFound)
-				return false;
-		}
-
-		return true;
-	}
-
-	std::vector<const char*> RendererContext::GetRequiredExtensions()
-	{
-		EPPO_PROFILE_FUNCTION("RendererContext::GetRequiredExtensions");
-
-		uint32_t glfwExtensionCount{ 0 };
-		const char** glfwExtensions = glfwGetRequiredInstanceExtensions(&glfwExtensionCount);
-
-		std::vector<const char*> extensions(glfwExtensions, glfwExtensions + glfwExtensionCount);
-
-		if (VulkanConfig::EnableValidation)
-			extensions.push_back(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
-
-		return extensions;
 	}
 }
