--- EppoEngine/Source/Renderer/Vulkan.h
@@ -1,86 +1,0 @@
-#pragma once
-
-#include <vulkan/vulkan.h>
-#include <tracy/TracyVulkan.hpp>
-
-namespace Eppo
-{
-	struct VulkanConfig
-	{
-		#if defined(EPPO_DEBUG) || defined(EPPO_RELEASE)
-			inline static const bool EnableValidation = true;
-		#else
-			inline static const bool EnableValidation = false;
-		#endif
-
-			inline static const uint32_t MaxFramesInFlight = 2;
-			inline static const std::vector<const char*> ValidationLayers = { "VK_LAYER_KHRONOS_validation" };
-			inline static const std::vector<const char*> DeviceExtensions = { VK_KHR_SWAPCHAIN_EXTENSION_NAME };
-	};
-
-	#define VK_CHECK(func, msg) { if (func != VK_SUCCESS) EPPO_ERROR(msg); }
-
-	static VKAPI_ATTR VkBool32 VKAPI_CALL DebugCallback(
-		VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity,
-		VkDebugUtilsMessageTypeFlagsEXT messageType,
-		const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData,
-		void* pUserData)
-	{
-		/*
-			MessageSeverity:
-			VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT: Diagnostic message
-			VK_DEBUG_UTILS_MESSAGE_SEVERITY_INFO_BIT_EXT: Informational message like the creation of a resource
-			VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT: Message about behavior that is not necessarily an error, but very likely a bug in your application
-			VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT: Message about behavior that is invalid and may cause crashes
-
-			MessageType:
-			VK_DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT: Some event has happened that is unrelated to the specification or performance
-			VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT: Something has happened that violates the specification or indicates a possible mistake
-			VK_DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT: Potential non-optimal use of Vulkan
-		*/
-		switch (messageSeverity)
-		{
-		case VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT:
-		{
-			EPPO_TRACE(pCallbackData->pMessage);
-			break;
-		}
-
-		case VK_DEBUG_UTILS_MESSAGE_SEVERITY_INFO_BIT_EXT:
-		{
-			EPPO_INFO(pCallbackData->pMessage);
-			break;
-		}
-
-		case VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT:
-		{
-			EPPO_WARN(pCallbackData->pMessage);
-			break;
-		}
-
-		case VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT:
-		{
-			EPPO_ERROR("Vulkan Error: {}", pCallbackData->pMessage);
-			break;
-		}
-		}
-
-		return VK_FALSE;
-	}
-
-	static VkResult CreateDebugUtilsMessengerEXT(VkInstance instance, const VkDebugUtilsMessengerCreateInfoEXT* pCreateInfo, const VkAllocationCallbacks* pAllocator, VkDebugUtilsMessengerEXT* pDebugMessenger)
-	{
-		auto func = (PFN_vkCreateDebugUtilsMessengerEXT)vkGetInstanceProcAddr(instance, "vkCreateDebugUtilsMessengerEXT");
-		if (func != nullptr)
-			return func(instance, pCreateInfo, pAllocator, pDebugMessenger);
-
-		return VK_ERROR_EXTENSION_NOT_PRESENT;
-	}
-
-	static void DestroyDebugUtilsMessengerEXT(VkInstance instance, VkDebugUtilsMessengerEXT debugMessenger, const VkAllocationCallbacks* pAllocator)
-	{
-		auto func = (PFN_vkDestroyDebugUtilsMessengerEXT)vkGetInstanceProcAddr(instance, "vkDestroyDebugUtilsMessengerEXT");
-		if (func != nullptr)
-			func(instance, debugMessenger, pAllocator);
-	}
-}
